pyproject.toml
[project]
name = "magnetic-report"
version = "0.1.0"
description = "API da aplicação magnetic report"
authors = [
    {name = "Lucas Beserra",email = "lbeserra.under@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14,<4"
dependencies = [
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-multipart (==0.0.6)",
    "pillow (>=12.0.0,<13.0.0)",
    "pypdf2 (==3.0.1)",
    "reportlab (>=4.4.5,<5.0.0)",
    "email-validator (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
